In PCRE2.replace callbacks, pass null arguments for unmatched captures
The current design has a problem: PCRE2 only returns matches against
captures up to the last capture that matches. An optional capture may not
match, yet the overall match is still successful. This means that the number
of subpattern arguments passed to a PCRE2.replace callback can vary.

Fixing this requires adding the ability to jpcre2 to get the number of
captures in the pattern.

Use the new method getNumCaptures to set the size of the captures array, and
pad out the array with “null” values. As a result, the number of arguments
passed to the callback is always the same.

This reduces confusion, and restores sanity in TypeScript.