Make unset captures null, not the empty string
It is important to be able to differentiate an empty match from a non-match.

Unfortunately, JPCRE2 does not permit this as-is. I have submitted a PR to
fix it: https://github.com/jpcre2/jpcre2/pull/31

Since the JPCRE2 maintainer is happy to take some form of the patch, but has
not found time to deal with it, I have replaced the JPCRE2 repo with my fork
that contains the PR.

Finally, this commit slightly simplifies binding.gyp: it removes the -O3
settings, as binding-gyp uses this by default anyway now, and updates the
language version to C++17, as this is required for my patch to JPCRE2.